TCOMP-125 Move Avro <-> MetadataTable conversions to its own helper class.
